requirements.txt
# Core LLM providers
openai>=1.0.0
anthropic>=0.18.0
cohere>=5.0.0
google-generativeai>=0.5.0

# Tokenization and utilities
tiktoken>=0.5.0
python-dotenv>=1.0.0
python-slugify>=8.0.1
requests>=2.31.0
aiohttp>=3.9.0  # For async HTTP with connection pooling

# Dependency injection
dependency-injector>=4.41.0

# RAG and Vector Search dependencies
sentence-transformers>=2.2.2  # For semantic embeddings
faiss-cpu>=1.7.4  # For vector similarity search
numpy>=1.21.0  # For numerical operations

# Caching backends (optional)
redis>=5.0.0  # For Redis cache and queue

# Background processing (optional)
celery>=5.3.0  # For Celery task queue
rq>=1.15.0  # For RQ task queue

# Performance monitoring (optional)
psutil>=5.9.0  # System monitoring